+++ samples/kotlin-mpp/src/commonMain/kotlin/org/jlleitschuh/gradle/ktlint/sample/mpp/CommonInterface.kt
@@ -1,0 +1,6 @@
+package org.jlleitschuh.gradle.ktlint.sample.mpp
+
+interface CommonInterface {
+    fun init()
+    fun getName(): String
+}