+++ samples/kotlin-mpp/src/linuxMain/kotlin/org/jlleitschuh/gradle/ktlint/sample/mpp/LinuxCommon.kt
@@ -1,0 +1,9 @@
+package org.jlleitschuh.gradle.ktlint.sample.mpp
+
+class LinuxCommon : CommonInterface {
+    override fun init() {
+        println("Init in linux")
+    }
+
+    override fun getName(): String = "LinuxCommon"
+}